requirements.txt
google-adk
jira
pytest
python-dotenv
fastapi 
uvicorn